test(research): split PDF caption region coverage
